requirements.txt
sanic
sanic_session
sanic-aioorm
sanic-aredis
sanic_cors
aiohttp
pillow
itsdangerous
aredis
bcrypt
